requirements.txt
Django==2.2.4
psycopg2>=2.7,<3.0
djangorestframework>=3.10
markdown
django-filter
requests
beautifulsoup4
